+++ src/main.rs
@@ -15,8 +15,10 @@ extern crate tokio;
 extern crate blake2;
 extern crate clap;
 extern crate dirs;
+extern crate glob;
 extern crate pathdiff;
 extern crate semver;
+extern crate tempdir;
 extern crate typed_arena;
 
 #[macro_use]